chore(dx): Update boilerplate.py
